Update CartItem.jsx
Task 3: CartItems component
Next, you will complete the development of the CartItem.jsxcomponent which displays the items held in the shopping cart. This component has a number of functionalities that you find in a typical shopping cart:

Calculate the total for all items in the cart.
Calculate the subtotal for each plant type in the cart.
Continue shopping
Increment and decrement the number of each plant type in the cart
Remove (delete) a plant type from the cart altogether.